pyproject.toml
[tool.poetry]
name = "sanx"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
pyarrow = "^15.0.1"
sqlalchemy = "^2.0.28"
dash = "^2.16.1"
dash-bootstrap-components = "^1.5.0"
sqlmodel = "^0.0.16"
plotly = "^5.20.0"
pillow = "^10.2.0"
cachetools = "^5.3.3"
dash-extensions = "^1.0.14"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
